Update azuredeploy.json to include dnsZone default values
Parameters were added for existingDnsZones and dnsZoneNames but with no default value. When utilizing the "Deploy to Azure" workflow these fields are required to be completed but it is unclear what values should be there as a baseline to users who aren't familiar. Adding in the defaults simplifies the process while still allowing modifications if necessary.